requirements.txt
google-cloud-bigquery==3.0.1
plotly==4.14.3
